Move get_root_path back to flask.helpers
Many libraries use it so the move is somewhat disruptive.
